pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "paper-assistant"
version = "0.1.0"
description = "AI-powered ML research paper summarizer with podcast generation"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"

dependencies = [
    "anthropic>=0.39.0",
    "click>=8.1.0",
    "httpx>=0.27.0",
    "pymupdf>=1.24.0",
    "pymupdf4llm>=0.0.10",
    "edge-tts>=6.1.0",
    "feedgen>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "jinja2>=3.1.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.0",
]

[project.scripts]
paper-assist = "paper_assistant.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/paper_assistant"]

[tool.ruff]
line-length = 100
target-version = "py310"
